+++ modules/Render/RenderOpenGL/src/Platform/GLPlatformGLX.c
@@ -374,7 +374,7 @@ bool dsGetGLSurfaceSize(uint32_t* outWidth, uint32_t* outHeight, void* display,
 	return true;
 }
 
-void dsSwapGLBuffers(void* display, dsRenderSurface** renderSurfaces, size_t count, bool vsync)
+void dsSwapGLBuffers(void* display, dsRenderSurface** renderSurfaces, uint32_t count, bool vsync)
 {
 	// vsync on the first surface to avoid waiting for multiple swaps with multiple surfaces.
 	// Allow vsync for all surfaces if swap groups are supported.
